+++ src/tags/TagsTable.tsx
@@ -1,6 +1,6 @@
 import { FC, useEffect, useRef } from 'react';
 import { splitEvery } from 'ramda';
-import { RouteChildrenProps } from 'react-router';
+import { useLocation } from 'react-router-dom';
 import { SimpleCard } from '../utils/SimpleCard';
 import SimplePaginator from '../common/SimplePaginator';
 import { useQueryState } from '../utils/helpers/hooks';
@@ -18,10 +18,11 @@ export interface TagsTableProps extends TagsListChildrenProps {
 const TAGS_PER_PAGE = 20; // TODO Allow customizing this value in settings
 
 export const TagsTable = (TagsTableRow: FC<TagsTableRowProps>) => (
-  { sortedTags, selectedServer, location, orderByColumn, currentOrder }: TagsTableProps & RouteChildrenProps,
+  { sortedTags, selectedServer, orderByColumn, currentOrder }: TagsTableProps,
 ) => {
   const isFirstLoad = useRef(true);
-  const { page: pageFromQuery = 1 } = parseQuery<{ page?: number | string }>(location.search);
+  const { search } = useLocation();
+  const { page: pageFromQuery = 1 } = parseQuery<{ page?: number | string }>(search);
   const [ page, setPage ] = useQueryState<number>('page', Number(pageFromQuery));
   const pages = splitEvery(TAGS_PER_PAGE, sortedTags);
   const showPaginator = pages.length > 1;